Move user's controllers to a single file
